+++ package.json
@@ -25,7 +25,8 @@
         "mongoose": "^8.0.4",
         "multer": "^1.4.5-lts.1",
         "multer-storage-cloudinary": "^4.0.0",
-        "nodemailer": "^6.9.11"
+        "nodemailer": "^6.9.11",
+        "twilio": "^4.23.0"
     },
     "devDependencies": {
         "nodemon": "^3.0.2"
